Added Dialog for Spells, Changed Combat Dialog, added Bonus Dice field
Combat rolls now use a Dropdown menü to select target defense level to calculate dn of the roll. If the token target function is used the dropdown is disabled.
Added a field for target armour points. This field is populated with enemy armour and disabled when the token target function is used.
Added Field for Bonus Dice to common-, spell- and combat-roll dialogs. Talents, Spells and Miracles can grant additional dice to a test