Changed display of additional Successes for skill and combat rolls
In Combat test complexity to count additional successes is basically zero since every success is counted as an additional success
Normal Skill rolls however have a variable complexity which needs to be substracted from to total number of successes to get the additional ones. 
To achieve this a simple boolean flag is added to indicate if a roll is made for melee / ranged combat  or a normal skill roll.